backend/pyproject.toml
[project]
name = "travel-planner-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Mykola Baranovskiy",email = "baranovskiynick@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "dj-database-url (>=3.1.2,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "drf-yasg (>=1.21.14,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
